[update] Pi Comms Rx integration
