+++ FIRMWARE/LFW513__RLOOP__POWER_NODE/.cproject
@@ -76,7 +76,7 @@
 						</toolChain>
 					</folderInfo>
 					<sourceEntries>
-						<entry excluding="RLOOP/LCCM656__RLOOP__PI_COMMS/RX/rI2CRX.c|RM4/LCCM221__RM4__SELFTEST/rm4_selftest.c|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/rm4_pcr.c|RM4/LCCM222__RM4__PIN_MUX/rm4_pinmux.c|MULTICORE/LCCM517__MULTICORE__SAFETOOL_BRANCH/safetool_branch.c|MULTICORE/LCCM517__MULTICORE__SAFETOOL_BRANCH/STRINGS|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/fault_tree.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/numerical.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TOLERANCE|RM4/LCCM282__RM4__SCI/rm4_sci.c|RM4/LCCM282__RM4__SCI/LOOPBACK|RM4/LCCM282__RM4__SCI/INTERRUPTS|RM4/LCCM282__RM4__SCI/HELPERS|RM4/LCCM280__RM4__MIBSPI_135/rm4_mibspi135.c|RM4/LCCM280__RM4__MIBSPI_135/TX|RM4/LCCM280__RM4__MIBSPI_135/SPEED|RM4/LCCM280__RM4__MIBSPI_135/PINS|RM4/LCCM280__RM4__MIBSPI_135/LOOPBACK|RM4/LCCM280__RM4__MIBSPI_135/INTERRUPTS|RM4/LCCM280__RM4__MIBSPI_135/ERRORS|RM4/LCCM229__RM4__DMA/rm4_dma.c|RM4/LCCM229__RM4__DMA/SUSPEND|RM4/LCCM229__RM4__DMA/NOTIFICATIONS|RM4/LCCM229__RM4__DMA/INTERRUPTS|RM4/LCCM228__RM4__DELAYS/rm4_delays.c|RM4/LCCM227__RM4__BIST/rm4_bist.c|RM4/LCCM225__RM4__VECTORED_ISR_MGR/rm4_vim.c|RM4/LCCM224__RM4__EFUSE/rm4_efuse.c|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/rm4_esm.c|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/rm4_esm__user_template.c|RM4/LCCM219__RM4__SYSTEM/rm4_system.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__startup.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__oscillator.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__aborts.c|RM4/LCCM215__RM4__I2C/rm4_i2c.c|RM4/LCCM215__RM4__I2C/rm4_i2c__user.c|RM4/LCCM215__RM4__I2C/WAIT|RM4/LCCM215__RM4__I2C/TIMEOUT|RM4/LCCM215__RM4__I2C/PINS|RM4/LCCM135__RM4__FLASH/rm4_flash.c|RM4/LCCM135__RM4__FLASH/PROGRAM|RM4/LCCM133__RM4__GIO/rm4_gio.c|RM4/LCCM133__RM4__GIO/rm4_gio__interrupts.c|RM4/LCCM133__RM4__GIO/PULLS|RM4/LCCM215__RM4__I2C/UNIT_TEST|RM4/LCCM280__RM4__MIBSPI_135/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/VECTORS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TRIG_TABLES|MULTICORE/LCCM118__MULTICORE__NUMERICAL/STRING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POWER|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POINTS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MATRIX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MAC_ADDX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/LINEAR|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FLOAT|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FILTERING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/CIRCLE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ARCTAN2|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ABS|RM4/LCCM229__RM4__DMA/UNIT_TEST|RM4/LCCM228__RM4__DELAYS/UNIT_TEST|RM4/LCCM227__RM4__BIST/UNIT_TEST|RM4/LCCM225__RM4__VECTORED_ISR_MGR/UNIT_TEST|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/UNIT_TEST|RM4/LCCM222__RM4__PIN_MUX/UNIT_TEST|RM4/LCCM221__RM4__SELFTEST/UNIT_TEST|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/UNIT_TEST|RM4/LCCM135__RM4__FLASH/UNIT_TEST|RM4/LCCM133__RM4__GIO/UNIT_TEST|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST/FUNCTION_ENTRY_TESTS/LCCM118_TS_034.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST/FUNCTION_ENTRY_TESTS/LCCM118_TS_033.c|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_C28x_FPU32.lib|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_CortexR4_LE.lib" flags="VALUE_WORKSPACE_PATH|RESOLVED" kind="sourcePath" name=""/>
+						<entry excluding="RM4/LCCM221__RM4__SELFTEST/rm4_selftest.c|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/rm4_pcr.c|RM4/LCCM222__RM4__PIN_MUX/rm4_pinmux.c|MULTICORE/LCCM517__MULTICORE__SAFETOOL_BRANCH/safetool_branch.c|MULTICORE/LCCM517__MULTICORE__SAFETOOL_BRANCH/STRINGS|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/fault_tree.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/numerical.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TOLERANCE|RM4/LCCM282__RM4__SCI/rm4_sci.c|RM4/LCCM282__RM4__SCI/LOOPBACK|RM4/LCCM282__RM4__SCI/INTERRUPTS|RM4/LCCM282__RM4__SCI/HELPERS|RM4/LCCM280__RM4__MIBSPI_135/rm4_mibspi135.c|RM4/LCCM280__RM4__MIBSPI_135/TX|RM4/LCCM280__RM4__MIBSPI_135/SPEED|RM4/LCCM280__RM4__MIBSPI_135/PINS|RM4/LCCM280__RM4__MIBSPI_135/LOOPBACK|RM4/LCCM280__RM4__MIBSPI_135/INTERRUPTS|RM4/LCCM280__RM4__MIBSPI_135/ERRORS|RM4/LCCM229__RM4__DMA/rm4_dma.c|RM4/LCCM229__RM4__DMA/SUSPEND|RM4/LCCM229__RM4__DMA/NOTIFICATIONS|RM4/LCCM229__RM4__DMA/INTERRUPTS|RM4/LCCM228__RM4__DELAYS/rm4_delays.c|RM4/LCCM227__RM4__BIST/rm4_bist.c|RM4/LCCM225__RM4__VECTORED_ISR_MGR/rm4_vim.c|RM4/LCCM224__RM4__EFUSE/rm4_efuse.c|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/rm4_esm.c|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/rm4_esm__user_template.c|RM4/LCCM219__RM4__SYSTEM/rm4_system.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__startup.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__oscillator.c|RM4/LCCM219__RM4__SYSTEM/rm4_system__aborts.c|RM4/LCCM215__RM4__I2C/rm4_i2c.c|RM4/LCCM215__RM4__I2C/rm4_i2c__user.c|RM4/LCCM215__RM4__I2C/WAIT|RM4/LCCM215__RM4__I2C/TIMEOUT|RM4/LCCM215__RM4__I2C/PINS|RM4/LCCM135__RM4__FLASH/rm4_flash.c|RM4/LCCM135__RM4__FLASH/PROGRAM|RM4/LCCM133__RM4__GIO/rm4_gio.c|RM4/LCCM133__RM4__GIO/rm4_gio__interrupts.c|RM4/LCCM133__RM4__GIO/PULLS|RM4/LCCM215__RM4__I2C/UNIT_TEST|RM4/LCCM280__RM4__MIBSPI_135/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/VECTORS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TRIG_TABLES|MULTICORE/LCCM118__MULTICORE__NUMERICAL/STRING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POWER|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POINTS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MATRIX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MAC_ADDX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/LINEAR|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FLOAT|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FILTERING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/CIRCLE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ARCTAN2|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ABS|RM4/LCCM229__RM4__DMA/UNIT_TEST|RM4/LCCM228__RM4__DELAYS/UNIT_TEST|RM4/LCCM227__RM4__BIST/UNIT_TEST|RM4/LCCM225__RM4__VECTORED_ISR_MGR/UNIT_TEST|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/UNIT_TEST|RM4/LCCM222__RM4__PIN_MUX/UNIT_TEST|RM4/LCCM221__RM4__SELFTEST/UNIT_TEST|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/UNIT_TEST|RM4/LCCM135__RM4__FLASH/UNIT_TEST|RM4/LCCM133__RM4__GIO/UNIT_TEST|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST/FUNCTION_ENTRY_TESTS/LCCM118_TS_034.c|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST/FUNCTION_ENTRY_TESTS/LCCM118_TS_033.c|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_C28x_FPU32.lib|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_CortexR4_LE.lib" flags="VALUE_WORKSPACE_PATH|RESOLVED" kind="sourcePath" name=""/>
 					</sourceEntries>
 				</configuration>
 			</storageModule>
