+++ FIRMWARE/LFW513__RLOOP__POWER_NODE/SOURCE/MAIN/localdef.h
@@ -17,6 +17,8 @@ RLOOP - PI COMMUNICATIONS MODULE
 		#define C_LOCALDEF__LCCM656__USE_ON_RM4								(1U)
 		#define C_LOCALDEF__LCCM656__USE_ON_WIN32							(0U)
 
+		/** enable the receiver side? */
+		#define C_LOCALDEF__LCCM656__ENABLE_RX								(0U)
 
 		/** Testing Options */
 		#define C_LOCALDEF__LCCM656__ENABLE_TEST_SPEC						(0U)
@@ -89,22 +91,6 @@ MS5607 - PRESSURE SENSOR
 		#include <MULTICORE/LCCM648__MULTICORE__MS5607/ms5607.h>
 
 	#endif //#if C_LOCALDEF__LCCM648__ENABLE_THIS_MODULE == 1U
-
-
-/*******************************************************************************
-TSYS01 - DIGITAL TEMP SENSOR
-*******************************************************************************/
-	#define C_LOCALDEF__LCCM647__ENABLE_THIS_MODULE							(1U)
-	#if C_LOCALDEF__LCCM647__ENABLE_THIS_MODULE == 1U
-
-
-		/** Testing Options */
-		#define C_LOCALDEF__LCCM647__ENABLE_TEST_SPEC						(0U)
-
-		/** Main include file */
-		#include <MULTICORE/LCCM647__MULTICORE__TSYS01/tsys01.h>
-
-	#endif //#if C_LOCALDEF__LCCM647__ENABLE_THIS_MODULE == 1U
 
 /*******************************************************************************
 DS2482S - I2C TO 1-WIRE CONVERTER
