+++ FIRMWARE/PROJECT_CODE/LCCM656__RLOOP__PI_COMMS/pi_comms.h
@@ -21,6 +21,42 @@
 		#define RPOD_I2C_FRAME_END 				0xD8U
 		
 
+		#define I2C_BUFFER_SIZE 3000
+		#define I2C_CONTROL_CHAR 0xD5
+		#define I2C_FRAME_START 0xD0
+		#define I2C_PARAMETER_START 0xD3
+		#define I2C_FRAME_END 0xD8
+
+		struct rI2CRX_decParam
+		{
+			Luint16 index;
+			Luint8 type;
+			Luint8 length;
+			void *val;
+		}rI2CRX_decParam_t;
+
+		extern void (*rI2CRX_frameRXBeginCB) ();
+		extern void (*rI2CRX_recvDecParamCB) (struct rI2CRX_decParam decParam);
+		extern void (*rI2CRX_frameRXEndCB) ();
+
+
+		enum rI2C_paramTypes
+		{
+			rI2C_INT8 = 0x11,
+			rI2C_UINT8 = 0x12,
+			rI2C_INT16 = 0x21,
+			rI2C_UINT16 = 0x22,
+			rI2C_INT32 = 0x41,
+			rI2C_UINT32 = 0x42,
+			rI2C_INT64 = 0x81,
+			rI2C_UINT64 = 0x82,
+			rI2C_FLOAT = 0x43,
+			rI2C_DOUBLE = 0x83
+		};
+
+		void rI2CRX_begin();
+		void rI2CRX_receiveBytes(Luint8* data, Luint16 length);
+
 		/*******************************************************************************
 		Structures
 		*******************************************************************************/
