+++ FIRMWARE/VERIFICATION/LFW540__VERIFY__LCCM656_TE_000/.cproject
@@ -77,7 +77,7 @@
 						</toolChain>
 					</folderInfo>
 					<sourceEntries>
-						<entry excluding="RLOOP/LCCM656__RLOOP__PI_COMMS/RX/rI2CRX.c|RM4/LCCM229__RM4__DMA/UNIT_TEST|RM4/LCCM228__RM4__DELAYS/UNIT_TEST|RM4/LCCM227__RM4__BIST/UNIT_TEST|RM4/LCCM225__RM4__VECTORED_ISR_MGR/UNIT_TEST|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/UNIT_TEST|RM4/LCCM222__RM4__PIN_MUX/UNIT_TEST|RM4/LCCM221__RM4__SELFTEST/UNIT_TEST|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/UNIT_TEST|RM4/LCCM215__RM4__I2C/UNIT_TEST|RM4/LCCM135__RM4__FLASH/UNIT_TEST|RM4/LCCM133__RM4__GIO/UNIT_TEST|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/VECTORS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TRIG_TABLES|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TOLERANCE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/STRING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POWER|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POINTS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MATRIX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MAC_ADDX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/LINEAR|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FLOAT|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FILTERING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/CIRCLE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ARCTAN2|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ABS|SOURCE/LCCM109__MULTICORE__SDIO/TESTS|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_C28x_FPU32.lib|SOURCE/Copy170314 of LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_CortexR4_LE.lib|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/SecurityCopy|SOURCE/LCCM127__RM4__USB/BUFFERS/rm4_usb__ringbuf.c|SOURCE/Copy190314 of LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM127__RM4__USB/CONTROL_LAYER|SOURCE/LCCM127__RM4__USB/BUFFERS/rm4_usb__buffer.c|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/RM48_usb_dmsc_disk - Copy.c|SOURCE/LCCM405__MULTICORE__MSD_USB_LAYER/SAFETOOL|SOURCE/LCCM224__RM4__EFUSE/SAFETOOL|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/deprecated|SOURCE/LCCM241__MULTICORE__STD_USB_LAYERxx|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/SAFETOOL" flags="VALUE_WORKSPACE_PATH|RESOLVED" kind="sourcePath" name=""/>
+						<entry excluding="RM4/LCCM229__RM4__DMA/UNIT_TEST|RM4/LCCM228__RM4__DELAYS/UNIT_TEST|RM4/LCCM227__RM4__BIST/UNIT_TEST|RM4/LCCM225__RM4__VECTORED_ISR_MGR/UNIT_TEST|RM4/LCCM223__RM4__PERIPH_CENTRAL_RESOURCE/UNIT_TEST|RM4/LCCM222__RM4__PIN_MUX/UNIT_TEST|RM4/LCCM221__RM4__SELFTEST/UNIT_TEST|RM4/LCCM220__RM4__ERROR_SIGNAL_MODULE/UNIT_TEST|RM4/LCCM215__RM4__I2C/UNIT_TEST|RM4/LCCM135__RM4__FLASH/UNIT_TEST|RM4/LCCM133__RM4__GIO/UNIT_TEST|MULTICORE/LCCM284__MULTICORE__FAULT_TREE/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/VECTORS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/UNIT_TEST|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TRIG_TABLES|MULTICORE/LCCM118__MULTICORE__NUMERICAL/TOLERANCE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/STRING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POWER|MULTICORE/LCCM118__MULTICORE__NUMERICAL/POINTS|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MATRIX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/MAC_ADDX|MULTICORE/LCCM118__MULTICORE__NUMERICAL/LINEAR|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FLOAT|MULTICORE/LCCM118__MULTICORE__NUMERICAL/FILTERING|MULTICORE/LCCM118__MULTICORE__NUMERICAL/CIRCLE|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ARCTAN2|MULTICORE/LCCM118__MULTICORE__NUMERICAL/ABS|SOURCE/LCCM109__MULTICORE__SDIO/TESTS|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_C28x_FPU32.lib|SOURCE/Copy170314 of LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM135__RM4__FLASH/F021 Flash API/v1.51.0/Libraries/F021_API_CortexR4_LE.lib|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/SecurityCopy|SOURCE/LCCM127__RM4__USB/BUFFERS/rm4_usb__ringbuf.c|SOURCE/Copy190314 of LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM127__RM4__USB/CONTROL_LAYER|SOURCE/LCCM127__RM4__USB/BUFFERS/rm4_usb__buffer.c|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/RM48_usb_dmsc_disk - Copy.c|SOURCE/LCCM405__MULTICORE__MSD_USB_LAYER/SAFETOOL|SOURCE/LCCM224__RM4__EFUSE/SAFETOOL|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/deprecated|SOURCE/LCCM241__MULTICORE__STD_USB_LAYERxx|SOURCE/LCCM241__MULTICORE__STD_USB_LAYER/SAFETOOL" flags="VALUE_WORKSPACE_PATH|RESOLVED" kind="sourcePath" name=""/>
 					</sourceEntries>
 				</configuration>
 			</storageModule>
